+++ index.html
@@ -12,7 +12,7 @@
     <div class="container">
       <div class="main-header">
         <header class="title" onClick="window.location.reload()">
-          <img src="./icons/Symbol.svg" class="news-logo" />
+          <img src="./asset/icons/Symbol.svg" class="news-logo" />
           <h1 class="display-bold24">뉴스스탠드</h1>
         </header>
         <time class="today display-medium16">2023. 02. 10. 금요일</time>
