+++ main.js
@@ -1,5 +1,5 @@
 import { NewArticlePrint } from "./module/components/NewArticle.js";
-import { printGrid } from "./module/GridView.js";
+import { printGrid } from "./module/view/GridView.js";
 import SelectViewStyle from "./module/ViewStyle.js";
 
 function initDate() {
